Refactor modal and panel navigation.
