+++ src/components/panels/CardListPanel.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useMemo } from 'react';
-import { Identity, NamedIdentity } from '../Types';
+import { Identity, NamedIdentity } from '../../Types';
 import { DocumentCardPreview } from 'office-ui-fabric-react';
-import { CardsContext, ImagesContext } from '../Contexts';
+import { CardsContext, ImagesContext } from '../../Contexts';
 import { ItemListPanel } from './ItemListPanel';
 
 type Props = {
